+++ beaker/services/experiment.py
@@ -7,6 +7,8 @@ from ..exceptions import *
 from .service_client import ServiceClient
 
 if TYPE_CHECKING:
+    from datetime import datetime, timedelta
+
     from rich.progress import TaskID
 
 
@@ -214,6 +216,7 @@ class ExperimentClient(ServiceClient):
         experiment: Union[str, Experiment],
         task: Optional[Union[str, Task]] = None,
         quiet: bool = False,
+        since: Optional[Union[str, "datetime", "timedelta"]] = None,
     ) -> Generator[bytes, None, None]:
         """
         Download the logs for an experiment.
@@ -232,6 +235,10 @@ class ExperimentClient(ServiceClient):
         :param task: The task ID, name, or object of a specific task from the Beaker experiment
             to fetch logs for. Required if there are multiple tasks in the experiment.
         :param quiet: If ``True``, progress won't be displayed.
+        :param since: Only show logs since a particular time. Could be a UTC `datetime` object,
+            (naive datetimes will be treated as UTC), a timestamp in the form of RFC 3339
+            (e.g. "2013-01-02T13:23:37Z"), or a relative time
+            (e.g. a timedelta or a string like "42m").
 
         :raises ValueError: The experiment has no tasks or jobs, or the experiment has multiple tasks but
             ``task`` is not specified.
@@ -250,7 +257,7 @@ class ExperimentClient(ServiceClient):
                     f"Experiment {exp.id} has no jobs for task "
                     f"'{task if isinstance(task, str) else task.display_name}'"
                 )
-        return self.beaker.job.logs(job.id, quiet=quiet)
+        return self.beaker.job.logs(job.id, quiet=quiet, since=since)
 
     def metrics(
         self, experiment: Union[str, Experiment], task: Optional[Union[str, Task]] = None